requirements.txt
amqp
asgiref
async-timeout
attrs
autobahn
Automat
billiard
celery
certifi
cffi
channels
channels-redis
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
constantly
cryptography
daphne
Django
django-cors-headers
django-environ
django-filter
django_debug_toolbar
djangorestframework
djangorestframework-simplejwt
environ
exceptiongroup
hyperlink
idna==3.4
incremental
iniconfig
kombu
msgpack
packaging
pluggy
prompt-toolkit
psycopg2-binary
pyasn1
pyasn1-modules
pycparser
PyJWT
pyOpenSSL
pytest
pytest-django
pytz
redis
requests
service-identity
six
sqlparse
tomli
Twisted
txaio
typing_extensions
urllib3
vine
wcwidth
zope.interface
aws-cdk.aws-codestar-alpha
boto3
